BIG db fixed

+++ src/main/resources/db/changelog/changes/create-table-cars.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS `cars`  (
                          `model` VARCHAR(255) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
                          `brand` VARCHAR(255) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
                          `type` VARCHAR(255) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
-                         `inventory` BIGINT(0) UNSIGNED NOT NULL,
+                         `inventory` INTEGER UNSIGNED NOT NULL,
                          `daily_fee` DECIMAL(5, 2) UNSIGNED NOT NULL,
                          PRIMARY KEY (`id`) USING BTREE
 ) ENGINE = InnoDB CHARACTER SET = utf8 COLLATE = utf8_general_ci ROW_FORMAT = Dynamic;
